ADD: updated setups after installig services
